os-26 move logos to components
