+++ packages/components/src/index.js
@@ -1,0 +1,5 @@
+export { default as Borealis } from './logos/Borealis';
+export { default as Ecoplast } from './logos/Ecoplast';
+export { default as LAT } from './logos/LAT';
+export { default as MtmPlastics } from './logos/MtmPlastics';
+export { default as Rosier } from './logos/Rosier';
