+++ packages/components/src/logos/Ecoplast.jsx
