+++ packages/components/src/logos/index.js
