+++ packages/components/src/logos/utils.js
