--- packages/logo/src/index.js
@@ -1,5 +1,0 @@
-export { default as Borealis } from './brands/Borealis';
-export { default as Ecoplast } from './brands/Ecoplast';
-export { default as LAT } from './brands/LAT';
-export { default as MtmPlastics } from './brands/MtmPlastics';
-export { default as Rosier } from './brands/Rosier';
